+++ src/refactoring/components/admin/product/NewProductForm.tsx
@@ -1,28 +1,19 @@
-import { useState } from "react";
 import { Product } from "../../../../types";
+import { useNewProduct } from "../../../hooks/admin/useNewProduct";
 
 interface NewProductFormProps {
   onProductAdd: (product: Product) => void;
 }
 
-const initProductState = {
-  name: "",
-  price: 0,
-  stock: 0,
-  discounts: [],
-};
-
 const NewProductForm = ({ onProductAdd }: NewProductFormProps) => {
-  const [showNewProductForm, setShowNewProductForm] = useState(false);
-  const [newProduct, setNewProduct] =
-    useState<Omit<Product, "id">>(initProductState);
+  const {
+    showNewProductForm,
+    setShowNewProductForm,
+    newProduct,
+    setNewProduct,
+    handleAddNewProduct,
+  } = useNewProduct(onProductAdd);
 
-  const handleAddNewProduct = () => {
-    const productWithId = { ...newProduct, id: Date.now().toString() };
-    onProductAdd(productWithId);
-    setNewProduct(initProductState);
-    setShowNewProductForm(false);
-  };
   return (
     <>
       <button
